Update the votes in the database too now
