+++ controllers/pollController.js
@@ -55,30 +55,10 @@ exports.votePoll = function(req, res) {
 	var id = req.params.id;
 
 	var selectedChoice = req.body['poll-choice'];
-
-	db.get(id, function(err, body) {
-		h.handleError(err);
-
-		var choices = body.choices.map(function(choice, index) {
-			return {
-				id: 'choice-' + index,
-				answer: choice.answer,
-				votedOn: selectedChoice === choice.id ? choice.votedOn + 1 : choice.votedOn,
-			}
-		});
-
-		var updatedPoll = {
-			question: body.question,
-			choices: choices,
-			'_rev': body._rev
-		}
-
-		db.insert(updatedPoll, body._id, function(err2, body2) {
-			h.handleError(err);
-
-			res.redirect('/poll/' + id);
-		})
-	})
+	
+	db.update(id, selectedChoice, function() {
+		res.redirect('/poll/' + id);
+	});
 }
